+++ test/com/piposaude/add_test/weeks_test.clj
@@ -1,11 +1,11 @@
-(ns com.piposaude.relative-date-test.weeks-test
+(ns com.piposaude.add-test.weeks-test
   (:require [clojure.test :refer :all]
-            [com.piposaude.calenjars :refer [relative-date-add]]
+            [com.piposaude.calenjars :as calendar]
             [tick.alpha.api :as t]))
 
-(deftest should-calculate-correct-date-when-relative-date-add-date-with-weeks
+(deftest should-calculate-correct-date-when-add-date-with-weeks
   (are [days expected]
-       (= expected (relative-date-add (t/date "2020-07-29") days :weeks))
+       (= expected (calendar/add (t/date "2020-07-29") days :weeks))
     0 (t/date "2020-07-29")
     1 (t/date "2020-08-05")
     11 (t/date "2020-10-14")
@@ -14,9 +14,9 @@
     -4 (t/date "2020-07-01")
     -5 (t/date "2020-06-24")))
 
-(deftest should-calculate-correct-date-when-relative-date-add-date-time-with-weeks
+(deftest should-calculate-correct-date-when-add-date-time-with-weeks
   (are [days expected]
-       (= expected (relative-date-add (t/date-time "2020-07-29T10:15:00") days :weeks))
+       (= expected (calendar/add (t/date-time "2020-07-29T10:15:00") days :weeks))
     0 (t/date-time "2020-07-29T10:15:00")
     1 (t/date-time "2020-08-05T10:15:00")
     11 (t/date-time "2020-10-14T10:15:00")