+++ test/com/piposaude/add_test/years_test.clj
@@ -1,11 +1,11 @@
-(ns com.piposaude.relative-date-test.years-test
+(ns com.piposaude.add-test.years-test
   (:require [clojure.test :refer :all]
-            [com.piposaude.calenjars :refer [relative-date-add]]
+            [com.piposaude.calenjars :as calendar]
             [tick.alpha.api :as t]))
 
-(deftest should-calculate-correct-date-when-relative-date-add-date-with-years
+(deftest should-calculate-correct-date-when-add-date-with-years
   (are [days expected]
-       (= expected (relative-date-add (t/date "2020-02-29") days :years))
+       (= expected (calendar/add (t/date "2020-02-29") days :years))
     0 (t/date "2020-02-29")
     1 (t/date "2021-02-28")
     11 (t/date "2031-02-28")
@@ -14,9 +14,9 @@
     -4 (t/date "2016-02-29")
     -5 (t/date "2015-02-28")))
 
-(deftest should-calculate-correct-date-when-relative-date-add-date-time-with-years
+(deftest should-calculate-correct-date-when-add-date-time-with-years
   (are [days expected]
-       (= expected (relative-date-add (t/date-time "2020-02-29T10:15:00") days :years))
+       (= expected (calendar/add (t/date-time "2020-02-29T10:15:00") days :years))
     0 (t/date-time "2020-02-29T10:15:00")
     1 (t/date-time "2021-02-28T10:15:00")
     11 (t/date-time "2031-02-28T10:15:00")