+++ CHANGELOG.md
@@ -3,6 +3,7 @@
 * UNRELEASED
   * Added support for nth-day-of-the-week holidays
   * Now supports observed holidays
+  * You can now specify a holiday to begin only after a certain year
 
 * 0.5.0 - Jan 7, 2022
   * Upgraded to tick 0.4.32