+++ CHANGELOG.md
@@ -1,6 +1,7 @@
 # Changes
 
 * UNRELEASED
+  * The main API function has been simplified greatly: `relative-date-add` is now just `add`
   * Added support for nth-day-of-the-week holidays
   * Now supports observed holidays, including UK Christmas/Boxing Day Monday/Tuesday rule
   * You can now specify a holiday to begin or end only after a certain year