+++ CHANGELOG.md
@@ -1,6 +1,6 @@
 # Changes
 
-* UNRELEASED
+* 0.7.0 - 7Aug22
   * Now supports singular versions of :days :weeks :months :years and :business-days
   * Cleaned up the namespace, moving internal fns somewhere else to avoid confusion